Exercise 02: gray code

+++ Exercise02/gray_code.py
@@ -1,22 +1,19 @@
 from numpy import uint32
 
 
-
 def gray_code(n: uint32) -> uint32:
     try:
-        return n ^ (n >> 1)
+        return uint32(n ^ (n >> 1))
     except Exception as e:
         print(f"Error: {e}")
-  
-  
+
+
 def main():
     try:
         n = uint32(input("Enter a number: "))
         print(f"Result: {gray_code(n)}")
     except Exception as e:
         print(f"Error: {e}")
-
-
 
 
 if __name__ == "__main__":
